backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
werkzeug==3.0.1
PyJWT==2.8.0